backend/requirements.txt
requests==2.26.0
django==2.2.19
djangorestframework==3.12.4
pytest-pythonpath==0.7.3
djangorestframework-simplejwt==4.7.2
PyJWT==2.1.0
psycopg2-binary==2.9.3
djoser==2.1.0
